+++ lib/OrdinaryDiffEqRKN/src/alg_utils.jl
@@ -1,0 +1,20 @@
+alg_extrapolates(alg::IRKN3) = true
+alg_extrapolates(alg::IRKN4) = true
+
+alg_order(alg::IRKN3) = 3
+alg_order(alg::Nystrom4) = 4
+alg_order(alg::FineRKN4) = 4
+alg_order(alg::FineRKN5) = 5
+alg_order(alg::Nystrom4VelocityIndependent) = 4
+alg_order(alg::IRKN4) = 4
+alg_order(alg::Nystrom5VelocityIndependent) = 5
+alg_order(alg::DPRKN4) = 4
+alg_order(alg::DPRKN5) = 5
+alg_order(alg::DPRKN6) = 6
+alg_order(alg::DPRKN6FM) = 6
+alg_order(alg::DPRKN8) = 8
+alg_order(alg::DPRKN12) = 12
+alg_order(alg::ERKN4) = 4
+alg_order(alg::ERKN5) = 5
+alg_order(alg::ERKN7) = 7
+alg_order(alg::RKN4) = 4
